Убирает лишние модули
