--- js/errors/utils/on-button-close-message.js
@@ -1,20 +1,0 @@
-import {documentBody} from '../../modal/variables.js';
-import {onDocumentKeyDown} from '../../modal/utils/on-document-key-down.js';
-import {uploadMessageError, uploadMessageSuccess} from '../variables.js';
-
-const onButtonCloseMessage = (evt) => {
-  const buttonSuccessClose = document.querySelector('.success');
-  const buttonErrorClose = document.querySelector('.error');
-  if (evt.target.className === 'success' || evt.target.className === 'success__button') {
-    documentBody.removeEventListener('keydown', onDocumentKeyDown);
-    buttonSuccessClose.removeEventListener('click', onButtonCloseMessage);
-    uploadMessageSuccess.remove();
-  }
-  if (evt.target.className === 'error' || evt.target.className === 'error__button') {
-    documentBody.removeEventListener('keydown', onDocumentKeyDown);
-    buttonErrorClose.removeEventListener('click', onButtonCloseMessage);
-    uploadMessageError.remove();
-  }
-};
-
-export {onButtonCloseMessage};
