--- js/filters/variables.js
@@ -1,3 +1,0 @@
-const filterButtons = document.querySelector('.img-filters');
-
-export {filterButtons};
